Add the http-request-builder example
This adds the same example as the http-request but using the builder
syntax.

Most of the code is the same and a fair degree of the change from macro
to builder syntax is trivial. Using `Suspense` with the builder syntax
is not covered in the documentation. This also shows off that structs
that derive `Prop` have public builders which may be required using the
builder syntax - such as with `SuspenseProps`.