Add the `prop:*` directive to the view macro
Adds the ability to set properties directly in the view macro that
accepts any value that can be coverted into the `JsValue`.

Supports both static and dynamic values in the same manner that
attributes do currently.

Updates some of the examples to use the new directive when the type can
be used directly as a property instead of cloning or serializing the
value.